+++ R/create_beauti_options_v2_6.R
@@ -25,7 +25,7 @@ create_beauti_options_v2_6 <- function(
   sequence_indent = 8,
   nucleotides_uppercase = FALSE,
   status = "",
-  namespace = beautier::get_default_beast_namespace(),
+  namespace = beautier::get_default_beast_namespace_v2_6(),
   required = ""
 ) {
   beautier::create_beauti_options(