+++ tests/testthat/test-create_beast2_beast_xml.R
@@ -27,7 +27,15 @@ test_that("use, v2.6.2", {
 
 test_that("use, v2.6, with status", {
 
-  beauti_options <- create_beauti_options_v2_6(status = "noAutoSetClockRate")
+  beauti_options <- create_beauti_options_v2_6(
+    namespace = paste0(
+      "beast.core:beast.evolution.alignment:",
+      "beast.evolution.tree.coalescent:beast.core.util:beast.evolution.nuc:",
+      "beast.evolution.operators:beast.evolution.sitemodel:",
+      "beast.evolution.substitutionmodel:beast.evolution.likelihood"
+    ),
+    status = "noAutoSetClockRate"
+  )
   created <- create_beast2_beast_xml(
     beauti_options
   )